+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/angular
 
+## 2.0.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider-utils@4.0.6
+  - @zenning/ai@6.0.14
+
+## 2.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider-utils@4.0.5
+  - @zenning/ai@6.0.13
+
 ## 2.0.12
 
 ### Patch Changes