+++ packages/google/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/google
 
+## 3.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider@3.0.4
+  - @zenning/provider-utils@4.0.6
+
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider@3.0.3
+  - @zenning/provider-utils@4.0.5
+
 ## 3.0.4
 
 ### Patch Changes