+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/llamaindex
 
+## 2.0.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/ai@6.0.14
+
+## 2.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/ai@6.0.13
+
 ## 2.0.12
 
 ### Patch Changes