+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # ai-core-e2e-next-server
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/ai@6.0.14
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/ai@6.0.13
+
 ## 0.0.1
 
 ### Patch Changes