+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/valibot
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider-utils@4.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider-utils@4.0.5
+
 ## 2.0.4
 
 ### Patch Changes