+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/vue
 
+## 3.0.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider-utils@4.0.6
+  - @zenning/ai@6.0.14
+
+## 3.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider-utils@4.0.5
+  - @zenning/ai@6.0.13
+
 ## 3.0.12
 
 ### Patch Changes